requirements.txt
attr==0.3.1
attrs==19.3.0
matplotlib==3.4.3
nest_asyncio==1.5.1
numpy==1.19.5
PyYAML==6.0
scipy==1.7.1
tensorflow>=2.5.2
tensorflow_addons==0.15.0
tensorflow_datasets==4.4.0
tensorflow_federated==0.19.0
